pyproject.toml
[tool.poetry]
name = "s3-cheksum"
version = "0.1.0"
description = ""
authors = ["imadenugraha <imadenugraha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
boto3 = "^1.34.64"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
